Add Metal-native presentation backend (CAMetalLayer + async MetalFX)
New display.window.presentation_backend = auto | opengl | metal
(auto = Metal when the Vulkan renderer is configured). Under the
Metal backend the main window is SDL_WINDOW_METAL + CAMetalLayer
(ui/xemu-metal.m): the NV2A present frame is handed off as an
IOSurface/MTLTexture through a new get_present_frame renderer op,
composited with an MSL gamma blit, and ImGui renders via
imgui_impl_metal — removing the GL blit hop, glFlush, and CGL rebind
from every presented frame. The GL presentation path remains fully
selectable and is regression-tested.

UI: ui/xui/metal-helpers.mm is a Metal twin of gl-helpers (mask
decals, SDF logo port, render-to-texture panels, texture loaders,
GPU-readback screenshots/thumbnails); texture handles widen to
uintptr_t and feed ImTextureID on both backends. VGA fallback
uploads the pixman surface to an MTLTexture.

Async MetalFX (Metal backend only): the three waitUntilCompleted
stalls (1-5 ms/frame on the PFIFO thread) are replaced by an
MTLSharedEvent signal; the UI present pass encodes a GPU-side wait
on the frame's value. Outputs move to rings of 3 private
MTLTextures, lifting the macOS 26 1920px IOSurface cap — upscaling
now targets the panel (1280x960 -> 2880x2160 on 4K, verified) — and
giving frame interpolation genuinely distinct prev/cur inputs.

Pacing: vsync maps to CAMetalLayer.displaySyncEnabled; exclusive
fullscreen picks the highest refresh rate among native-resolution
modes. Renderer switches away from Vulkan under the Metal backend
prompt for a restart.

Build: Objective-C++ enabled on darwin (configure cross-file +
meson), imgui subproject metal backend enabled via packagefiles
overlay.

Validated: Azurik 90s Metal-backend soak (MetalFX spatial + 2x
interpolation at 2880x2160, no errors) and a GL-backend
(presentation_backend = 'opengl') regression run with CGL zero-copy
display intact.

+++ ui/xui/xemu-hud.h
@@ -25,6 +25,7 @@
 
 #include <SDL3/SDL.h>
 #include <epoxy/gl.h>
+#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -46,7 +47,7 @@ void xemu_hud_update(void);
 void xemu_hud_render(void);
 void xemu_hud_process_sdl_events(SDL_Event *event);
 void xemu_hud_should_capture_kbd_mouse(int *kbd, int *mouse);
-void xemu_hud_set_framebuffer_texture(GLuint tex, bool flip);
+void xemu_hud_set_framebuffer_texture(uintptr_t tex, bool flip);
 void xemu_set_framebuffer_texture_is_rect(bool is_rect);
 
 #ifdef __cplusplus
